Create wordle command that lets you start a game of wordle
Using /wordle, a player can start a game of wordle and start guessing words.
When a guess is entered it will be displayed on the embed if it was valid.
Timer is displayed on the embed adjusting every 10 seconds. Once the game is
over, the game data is saved to the database.